# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/products": {
      "get": {
        "operationId": "ProductController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Seitennummer (1-basiert)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Einträge pro Seite",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedProductsResponseDto"
                }
              }
            }
          }
        },
        "summary": "Produkte paginiert auflisten",
        "tags": [
          "products"
        ]
      },
      "post": {
        "operationId": "ProductController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validierungsfehler"
          }
        },
        "summary": "Neues Produkt anlegen",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "operationId": "ProductController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Produkt nicht gefunden"
          }
        },
        "summary": "Produkt per ID laden",
        "tags": [
          "products"
        ]
      },
      "patch": {
        "operationId": "ProductController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validierungsfehler"
          },
          "404": {
            "description": "Produkt nicht gefunden"
          }
        },
        "summary": "Produkt aktualisieren",
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Erfolgreich gelöscht"
          },
          "404": {
            "description": "Produkt nicht gefunden"
          }
        },
        "summary": "Produkt löschen",
        "tags": [
          "products"
        ]
      }
    },
    "/orders": {
      "get": {
        "operationId": "OrderController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "All orders",
        "tags": [
          "orders"
        ]
      },
      "post": {
        "operationId": "OrderController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Product not found"
          },
          "409": {
            "description": "Insufficient stock"
          }
        },
        "summary": "Place a new order",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "operationId": "OrderController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Order not found"
          }
        },
        "summary": "Order by ID",
        "tags": [
          "orders"
        ]
      }
    }
  },
  "info": {
    "title": "Inform Bestellsystem API",
    "description": "REST-API für Produkt- und Bestellverwaltung",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "ProductResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "name": {
            "type": "string",
            "example": "MacBook Pro 14\""
          },
          "price": {
            "type": "number",
            "example": 1999.99
          },
          "stock": {
            "type": "number",
            "example": 15
          }
        },
        "required": [
          "id",
          "name",
          "price",
          "stock"
        ]
      },
      "PaginatedProductsResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductResponseDto"
            }
          },
          "total": {
            "type": "number",
            "example": 42
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 10
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "limit"
        ]
      },
      "CreateProductRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "MacBook Pro 14\""
          },
          "price": {
            "type": "number",
            "example": 1999.99,
            "minimum": 0
          },
          "stock": {
            "type": "number",
            "example": 15,
            "minimum": 0
          }
        },
        "required": [
          "name",
          "price",
          "stock"
        ]
      },
      "UpdateProductRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "MacBook Pro 16\""
          },
          "price": {
            "type": "number",
            "example": 2499.99,
            "minimum": 0
          },
          "stock": {
            "type": "number",
            "example": 10,
            "minimum": 0
          }
        }
      },
      "OrderItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440010"
          },
          "productId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "quantity": {
            "type": "number",
            "example": 2
          },
          "unitPrice": {
            "type": "number",
            "example": 999.99
          },
          "subtotal": {
            "type": "number",
            "example": 1999.98
          }
        },
        "required": [
          "id",
          "productId",
          "quantity",
          "unitPrice",
          "subtotal"
        ]
      },
      "OrderResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440020"
          },
          "userId": {
            "type": "string",
            "example": "00000000-0000-0000-0000-000000000001"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemResponseDto"
            }
          },
          "totalPrice": {
            "type": "number",
            "example": 1999.98
          },
          "createdAt": {
            "type": "string",
            "example": "2026-08-16T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "userId",
          "items",
          "totalPrice",
          "createdAt"
        ]
      },
      "CreateOrderItemRequestDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "quantity": {
            "type": "number",
            "example": 2,
            "minimum": 1
          }
        },
        "required": [
          "productId",
          "quantity"
        ]
      },
      "CreateOrderRequestDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "00000000-0000-0000-0000-000000000001"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderItemRequestDto"
            }
          }
        },
        "required": [
          "userId",
          "items"
        ]
      }
    }
  }
}